Keep CMS-specific utilities in a separate subdirectory to allow for rolling back
