+++ starterkits/drupal/app/api/preview/route.ts
@@ -1,4 +1,4 @@
-import drupal from '@/util/drupal';
+import drupal from '@/util/drupal/drupal';
 import { draftMode } from 'next/headers';
 import { redirect } from 'next/navigation';
 import { NextRequest, NextResponse } from 'next/server';
