+++ starterkits/drupal/app/blog/[slug]/page.tsx
@@ -8,8 +8,8 @@ import {
   canShowEntity,
   entityExists,
   routeIsInternal,
-} from '@/util/dataIsEntityType';
-import query from '@/util/query';
+} from '@/util/drupal/dataIsEntityType';
+import query from '@/util/drupal/query';
 import { draftMode } from 'next/headers';
 import { notFound } from 'next/navigation';
 import ArticleFull from '../../_components/content/ArticleFull';
